requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiomysql==0.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.2.0
jinja2==3.1.3
xhtml2pdf==0.2.14
pypandoc==1.11.3
